refactor(api): relocate aiController
