+++ backend/api/src/routes/aiRoutes.ts
@@ -1,5 +1,5 @@
 import express from 'express';
-import * as aiController from '../controllers/shared/ai/aiController';
+import * as aiController from '../controllers/ai/aiController';
 import { protect } from '../middleware/authMiddleware';
 import { mutationLimiter } from '../middleware/rateLimiter';
 import { validateRequest } from '../middleware/validateRequest';
